The infrastructure tests hand numpy too, and a typed array keeps its dtype
Three test modules still built torch tensors for a harness that now takes
numpy, which the two fast suites I checked before pushing do not collect.
benchmark, comparison and the sequence infrastructure tests draw with
np.random.default_rng and write through numpy_view().

One of those failures was not the tests' fault. NpuTensor's constructor
defaults dtype to uint32 and does not read it off a typed ndarray, so
run_test's input buffer was reinterpreting bfloat16 as uint32 -- visible
only as a bandwidth figure 1.5x too high, since nothing else looks at the
buffer's dtype. It passes dtype=data.dtype now. from_torch had been hiding
this by converting explicitly.

Device-free is back to 45 failed / 1180 passed / 18 skipped, the failure
set identical to the pre-restyle baseline by name.

+++ iron/common/harness.py
@@ -205,7 +205,9 @@ def run_test(
                 produced[name] = buf
             else:
                 name, data = next(ins)
-                buf = tensor_class(data)
+                # dtype explicitly: the constructor's default is uint32, so a
+                # typed array would otherwise be reinterpreted, not adopted.
+                buf = tensor_class(data, dtype=data.dtype)
                 if b.direction == "inout":
                     produced[name] = buf
         except StopIteration:
